Add an edge client facade for edge features
The model registry will be the source of truth for the CLI
for models, versions and images. Add a facade that abstracts
the model registry and contains all the edge sepcific model
registry logic.

This will be a package that can be used by both the CLI
and the pipelines.